sass BEM and advanced socket

+++ client/src/hooks/useLocalStorageTracker.js
@@ -2,12 +2,9 @@ import { useEffect } from "react";
 
 // Custom hook for managing localStorage and detecting tab closures
 export const useLocalStorageTracker = () => {
-  //Quand j'ouvre une tab
-  //Dans un useEffect :
-  //Je crée un objet tabCounter  s'il n'existe pas ou j'incrémente de 1 s'il existe
-
   useEffect(() => {
     const handleTabClosing = () => {
+      if (localStorage.getItem("tabCounter") === null) return;
       if (localStorage.getItem("tabCounter") === "1") {
         localStorage.clear();
       } else {
@@ -17,11 +14,13 @@ export const useLocalStorageTracker = () => {
         );
       }
     };
-    const tabCounter = localStorage.getItem("tabCounter");
-    if (!tabCounter) {
+    if (localStorage.getItem("tabCounter") === null) {
       localStorage.setItem("tabCounter", "1");
     } else {
-      localStorage.setItem("tabCounter", (parseInt(tabCounter) + 1).toString());
+      localStorage.setItem(
+        "tabCounter",
+        (parseInt(localStorage.getItem("tabCounter")) + 1).toString()
+      );
     }
 
     window.addEventListener("beforeunload", handleTabClosing);
